單筆: 整理程式碼
