+++ src/main/java/com/example/pentaho/service/RedisService.java
@@ -286,111 +286,4 @@ public class RedisService {
         log.info("keySet:{}", keySet);
         return keySet;
     }
-
-
-    public Set<String> fuzzySearchMappingId(Address address) {
-        Map<String, List<String>> map = buildRegexMappingId(address);
-        List<String> newMappingId = map.get("newMappingId");
-        List<String> regex = map.get("regex");
-        log.info("因為地址不完整，組成新的 mappingId {}，以利模糊搜尋", newMappingId);
-        log.info("模糊搜尋正則表達式為:{}", regex);
-        Set<String> newMappingIdSet = new HashSet<>();
-        for (int i = 0; i < newMappingId.size(); i++) {
-            Set<String> mappingIdSet = findListByScan(newMappingId.get(i));
-            log.info("mappingIdSet:{}", mappingIdSet);
-            Pattern regexPattern = Pattern.compile(String.valueOf(regex.get(i)));
-            for (String newMapping : mappingIdSet) {
-                //因為redis的scan命令，無法搭配正則，限制*的位置只能有多少字元，所以要再用java把不符合的mappingId刪掉
-                Matcher matcher = regexPattern.matcher(newMapping);
-                //有符合的mappingId，才是真正要拿來處理比對代碼的mappingId
-                if (matcher.matches()) {
-                    newMappingIdSet.add(newMapping);
-                }
-            }
-        }
-        return newMappingIdSet;
-    }
-
-    private Map<String, List<String>> buildRegexMappingId(Address address) {
-        List<String> newMappingIdList = new ArrayList<>();
-        List<String> regexList = new ArrayList<>();
-        String segNum = address.getSegmentExistNumber();
-        StringBuilder newMappingId = new StringBuilder(); //組給redis的模糊搜尋mappingId ex.10010***9213552**95********
-        StringBuilder regex = new StringBuilder();   //組給java的模糊搜尋mappingId ex.10010020017029\d{18}95\d{19}000000\d{5}
-        //mappingCount陣列代表，COUNTY_CD要5位元，TOWN_CD要3位元，VILLAGE_CD要3位元，以此類推
-        //6,5,4,3,1 分別是NUM_FLR_1~NUM_FLR_5
-        int[] mappingCount = {5, 3, 3, 3, 7, 4, 7, 2, 6, 5, 4, 3, 1, 1, 5, 5};
-        int sum = 0;
-        for (int j = 0; j < address.getMappingIdList().size(); j++) {
-            for (int i = 0; i < segNum.length(); i++) {
-                //該欄位是1的情況(找的到的情況)
-                if ("1".equals(String.valueOf(segNum.charAt(i)))) {
-                    newMappingId.append(address.getMappingIdList().get(j).get(i));
-                    regex.append(address.getMappingIdList().get(j).get(i));
-                    sum = 0; //歸零
-                    //該欄位是0的情況(找不到的情況)
-                } else {
-                    String segAfter = "";
-                    String segBefore = "";
-                    //第一碼
-                    if (i == 0) {
-                        segAfter = String.valueOf(segNum.charAt(i + 1));
-                        //如果前後碼也是0的話，表示要相加
-                        if ("0".equals(segAfter)) {
-                            sum += mappingCount[i];
-                        }
-                        //後面那碼是1，表示不用相加了
-                        else if ("1".equals(segAfter)) {
-                            sum += mappingCount[i];
-                            regex.append("\\d{").append(sum).append("}");
-                            sum = 0; //歸零
-                        }
-                    }
-                    //不是最後一個
-                    else if (i != segNum.length() - 1) {
-                        segAfter = String.valueOf(segNum.charAt(i + 1));
-                        segBefore = String.valueOf(segNum.charAt(i - 1));
-                        //如果前後都是是1的話，不用相加
-                        if ("1".equals(segBefore) && "1".equals(segAfter)) {
-                            sum = 0; //歸零
-                            regex.append("\\d{").append(mappingCount[i]).append("}");
-                        }
-                        //如果前後碼也是0的話，表示要相加
-                        else if ("0".equals(segAfter)) {
-                            sum += mappingCount[i];
-                        }
-                        //後面那碼是1，表示不用相加了
-                        else if ("1".equals(segAfter)) {
-                            sum += mappingCount[i];
-                            regex.append("\\d{").append(sum).append("}");
-                            sum = 0; //歸零
-                        }
-                    }
-                    //是最後一個
-                    else {
-                        segBefore = String.valueOf(segNum.charAt(i - 1));
-                        //如果前面是0表示最後一碼也要加上去
-                        if ("0".equals(segBefore)) {
-                            sum += mappingCount[i];
-                            regex.append("\\d{").append(sum).append("}");
-                        } else {
-                            regex.append("\\d{").append(mappingCount[i]).append("}");
-                        }
-                        sum = 0; //歸零
-                    }
-                    newMappingId.append("*");
-                }
-            }
-            newMappingIdList.add(newMappingId.toString());
-            regexList.add(regex.toString());
-        }
-        Map<String, List<String>> map = new HashMap<>();
-        map.put("newMappingId", newMappingIdList); //帶有**的mappingId(redis模糊搜尋)
-        map.put("regex", regexList); //帶有正則的mappingId(java比對redis模糊搜尋出來的結果)
-        return map;
-    }
-
-
-
-
 }
